+++ devicemanagement/tests/TestDevMgmtApiSyncCmds.cpp
@@ -901,7 +901,7 @@ void TestDevMgmtApiSyncCmds::getModulePower(bool singleDevice) {
     DV_LOG(INFO) << "Service Request Completed for Device: " << deviceIdx;
 
     // Skip validation if loopback driver or SysEMU
-    if (targetInList({Target::Loopback, Target::SysEMU})) {
+    if (!targetInList({Target::Loopback, Target::SysEMU})) {
       // Note: Module power could vary. So there cannot be expected value for Module power in the test
       device_mgmt_api::module_power_t* module_power = (device_mgmt_api::module_power_t*)output_buff;
       auto power = (module_power->power >> 2) + (module_power->power & 0x03) * 0.25;
@@ -932,7 +932,7 @@ void TestDevMgmtApiSyncCmds::getModuleVoltage(bool singleDevice) {
     DV_LOG(INFO) << "Service Request Completed for Device: " << deviceIdx;
 
     // Skip validation if loopback driver or SysEMU
-    if (targetInList({Target::Loopback, Target::SysEMU})) {
+    if (!targetInList({Target::Loopback, Target::SysEMU})) {
       // Note: Module power could vary. So there cannot be expected value for Module power in the test
       device_mgmt_api::module_voltage_t* module_voltage = (device_mgmt_api::module_voltage_t*)output_buff;