+++ devicemanagement/tests/functional/FunctionalTestDevMgmtApiFirmwareMgmtCmds.cpp
@@ -37,7 +37,9 @@ TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, getMMErrorCount) {
 }
 
 TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, getFWBootstatus) {
-  if (targetInList({Target::FullBoot, Target::Silicon})) {
+  // TODO: SW-13807: Enable back on silicon after fix
+  // if (targetInList({Target::FullBoot, Target::Silicon})) {
+  if (targetInList({Target::FullBoot})) {
     getFWBootstatus(false /* Multiple devices */);
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";