+++ devicemanagement/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `setModuleFrequency` to `setAndGetModuleFrequency` because now this
   test sets the frequency and validates it by reading back those frequencies.
 - [SW-13583] Enabled getFWBootstatus test
+- [SW-13620] Add new command DM_CMD_GET_SP_STATS to the commandCodeTable
 ### Deprecated
 ### Removed
 ### Fixed